analyzer/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0
openslide-python>=1.2.0
openslide-bin>=4.0.0
